Moves TestSiteMap over to /results and renames it ResultsMap to keep it consistent with other components
